% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gd.R
\name{fit_gd}
\alias{fit_gd}
\title{Fit a linear regression using Gradient Descent}
\usage{
fit_gd(x, y, alpha = 0.01, iterations = 1000)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{alpha}{learning rate}

\item{iterations}{number of iterations}
}
\description{
Fit a linear regression using Gradient Descent
}
